requirements.txt
pytz>=2019.2
pyyaml>=5.1.2
matplotlib>=3.1.1
pandas>=0.25.0
numpy>=1.17.0
networkx>=2.3
torchvision>=0.4.0
scikit-learn>=0.21.3
scikit-image>=0.15.0
pytorch>=1.3.1
tensorboard>=1.14
opencv>=4.1.1.26
future

